+++ mobile_verifier/src/reward_share.rs
@@ -67,6 +67,15 @@ impl OwnerEmissions {
         OwnerEmissions(owner_emissions)
     }
 
+    #[allow(dead_code)]
+    pub fn total_emissions(&self) -> Mobile {
+        Mobile::from(
+            self.0
+                .values()
+                .fold(dec!(0), |acc, amt| acc + amt.into_inner()),
+        )
+    }
+
     pub fn into_inner(self) -> HashMap<PublicKey, Mobile> {
         self.0
     }
